feat: add BigInteger support for reading VarInt
This commit adds support for reading and returning `BigInteger` values when reading VarInts in the `BufferReader` class. The `readVarInt` method now returns a `BigInteger` instead of a `long`, and the necessary changes have been made to the dependent classes and tests.

+++ app/src/test/java/protobuf/magic/BufferReaderTest.java
@@ -1,9 +1,12 @@
 package protobuf.magic;
 
-import static org.junit.jupiter.api.Assertions.*;
-import static org.mockito.Mockito.*;
+import static org.junit.jupiter.api.Assertions.assertArrayEquals;
+import static org.junit.jupiter.api.Assertions.assertEquals;
+import static org.mockito.Mockito.spy;
 
 import java.lang.reflect.Field;
+import java.math.BigInteger;
+
 import org.junit.jupiter.api.Test;
 
 public class BufferReaderTest {
@@ -13,7 +16,7 @@ public class BufferReaderTest {
     byte[] buffer = BufferUtils.parseInput("8F01");
     BufferReader reader = new BufferReader(buffer);
 
-    long result = reader.readVarInt();
+    BigInteger result = reader.readVarInt();
     assertEquals(143, result);
 
     assertEquals(2, reader.getOffset());
